@@ -1,25 +1,22 @@
-﻿using System;
-using System.IO;
-using NagesenAsAService.Controllers;
+﻿using NagesenAsAService.Controllers;
 using NagesenAsAService.Models;
 using NUnit.Framework;
 
-namespace NagesenAsAService.Test
+namespace NagesenAsAService.Test;
+
+public class RoomsApiControllerTest
 {
-    public class RoomsApiControllerTest
+    [Test]
+    public void GetImageBytes_for_OGP_Image_Test()
     {
-        [Test]
-        public void GetImageBytes_for_OGP_Image_Test()
-        {
-            var testImagePath = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Fixtures", "UnavailableRoomNumber.jpg");
-            var picture = new Picture(DateTime.Now, () => File.ReadAllBytes(testImagePath));
+        var testImagePath = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Fixtures", "UnavailableRoomNumber.jpg");
+        var picture = new Picture(DateTime.Now, () => File.ReadAllBytes(testImagePath));
 
-            var actualBytes = RoomsApiController.GetImageBytes(picture, forOgpImage: true);
+        var actualBytes = RoomsApiController.GetImageBytes(picture, forOgpImage: true);
 
-            var resizedImagePath = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Fixtures", "resized.jpg");
-            var expectedBytes = File.ReadAllBytes(resizedImagePath);
+        var resizedImagePath = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Fixtures", "resized.jpg");
+        var expectedBytes = File.ReadAllBytes(resizedImagePath);
 
-            actualBytes.Is(expectedBytes);
-        }
+        actualBytes.Is(expectedBytes);
     }
 }
