@@ -1,16 +1,14 @@
-﻿using System;
-using Microsoft.AspNetCore.Http.Extensions;
+﻿using Microsoft.AspNetCore.Http.Extensions;
 using Microsoft.AspNetCore.Mvc;
 
-namespace NagesenAsAService.Extensions
+namespace NagesenAsAService.Extensions;
+
+public static class UrlHelperExtension
 {
-    public static class UrlHelperExtension
+    public static string AppUrl(this IUrlHelper urlHelper)
     {
-        public static string AppUrl(this IUrlHelper urlHelper)
-        {
-            var request = urlHelper.ActionContext.HttpContext.Request;
-            var requestUri = new Uri(request.GetDisplayUrl());
-            return requestUri.GetLeftPart(UriPartial.Scheme | UriPartial.Authority);
-        }
+        var request = urlHelper.ActionContext.HttpContext.Request;
+        var requestUri = new Uri(request.GetDisplayUrl());
+        return requestUri.GetLeftPart(UriPartial.Scheme | UriPartial.Authority);
     }
 }
