@@ -1,24 +1,22 @@
-﻿using System;
-using Newtonsoft.Json;
+﻿using Newtonsoft.Json;
 
-namespace NagesenAsAService.Models
+namespace NagesenAsAService.Models;
+
+public class reCAPTCHAVerificationResponse
 {
-    public class reCAPTCHAVerificationResponse
-    {
-        public bool Success { get; set; }
+    public bool Success { get; set; }
 
-        public double Score { get; set; }
+    public double Score { get; set; }
 
-        // timestamp of the challenge load (ISO format yyyy-MM-dd'T'HH:mm:ssZZ)
-        [JsonProperty("challenge_ts")]
-        public DateTimeOffset ChallengeTimestamp { get; set; }
+    // timestamp of the challenge load (ISO format yyyy-MM-dd'T'HH:mm:ssZZ)
+    [JsonProperty("challenge_ts")]
+    public DateTimeOffset ChallengeTimestamp { get; set; }
 
-        // the hostname of the site where the reCAPTCHA was solved
-        public string Hostname { get; set; }
+    // the hostname of the site where the reCAPTCHA was solved
+    public string Hostname { get; set; } = null!;
 
-        public string Action { get; set; }
+    public string Action { get; set; } = null!;
 
-        [JsonProperty("error-codes")]
-        public string[] ErrorCodes { get; set; } = new string[0];
-    }
+    [JsonProperty("error-codes")]
+    public string[] ErrorCodes { get; set; } = new string[0];
 }
