@@ -1,22 +1,19 @@
-﻿using System;
+﻿namespace NagesenAsAService.Models;
 
-namespace NagesenAsAService.Models
+public class Picture
 {
-    public class Picture
+    public DateTime LastModified { get; }
+
+    private Func<byte[]> ImageBytesGetter { get; }
+
+    public Picture(DateTime lastModified, Func<byte[]> getImageBytes)
     {
-        public DateTime LastModified { get; }
+        this.LastModified = lastModified;
+        this.ImageBytesGetter = getImageBytes;
+    }
 
-        private Func<byte[]> ImageBytesGetter { get; }
-
-        public Picture(DateTime lastModified, Func<byte[]> getImageBytes)
-        {
-            this.LastModified = lastModified;
-            this.ImageBytesGetter = getImageBytes;
-        }
-
-        public byte[] GetImageBytes()
-        {
-            return this.ImageBytesGetter();
-        }
+    public byte[] GetImageBytes()
+    {
+        return this.ImageBytesGetter();
     }
 }
