@@ -1,12 +1,9 @@
-﻿using System.Threading.Tasks;
+﻿namespace NagesenAsAService.Services.UrlShorter;
 
-namespace NagesenAsAService.Services.UrlShorter
+public class NoopUrlShorter : IUrlShorter
 {
-    public class NoopUrlShorter : IUrlShorter
+    public Task<string> ShortenUrlAsync(string url)
     {
-        public Task<string> ShortenUrlAsync(string url)
-        {
-            return Task.FromResult(url);
-        }
+        return Task.FromResult(url);
     }
 }
